+++ Test/Unit/Controller/Server/CancelTest.php
@@ -9,10 +9,12 @@ namespace Ebizmarts\SagePaySuite\Test\Unit\Controller\Server;
 use Magento\Framework\Encryption\EncryptorInterface;
 use Magento\Framework\Message\ManagerInterface;
 use Magento\Framework\TestFramework\Unit\Helper\ObjectManager as ObjectManagerHelper;
-
 use Ebizmarts\SagePaySuite\Controller\Server\Cancel;
 use Ebizmarts\SagePaySuite\Model\Config;
 use Ebizmarts\SagePaySuite\Model\Logger\Logger;
+use Ebizmarts\SagePaySuite\Model\ObjectLoader\OrderLoader;
+use Ebizmarts\SagePaySuite\Model\RecoverCart;
+use Magento\Checkout\Model\Cart;
 use Magento\Checkout\Model\Session;
 use Magento\Framework\App\Action\Action;
 use Magento\Framework\App\Action\Context;
@@ -20,13 +22,11 @@ use Magento\Framework\App\Request\Http as HttpRequest;
 use Magento\Framework\App\Response\Http as HttpResponse;
 use Magento\Framework\ObjectManager\ObjectManager;
 use Magento\Framework\UrlInterface;
-use Magento\Checkout\Model\Cart;
 use Magento\Quote\Model\Quote;
 use Magento\Quote\Model\QuoteIdMaskFactory;
 use Magento\Sales\Model\Order;
 use Magento\Sales\Model\OrderFactory;
 use Psr\Log\LoggerInterface;
-use Ebizmarts\SagePaySuite\Model\RecoverCart;
 
 class CancelTest extends \PHPUnit_Framework_TestCase
 {
@@ -143,25 +143,38 @@ class CancelTest extends \PHPUnit_Framework_TestCase
 
     public function testExecute()
     {
-        $this->request->expects($this->exactly(3))
+        $storeId = 1;
+        $quoteId = 69;
+        $encrypted = '0:2:Dwn8kCUk6nZU5B7b0Xn26uYQDeLUKBrD:S72utt9n585GrslZpDp+DRpW+8dpqiu/EiCHXwfEhS0=';
+        $encryptedOrderId = '0:3:Lq/5e1tdLdR19OaUuu1JTxD+7secLH91mWTNsT9c';
+        $orderId = 44;
+
+        $this->messageManager->expects($this->once())
+            ->method('addError')->willReturn($this->messageManager);
+
+        $this->request->expects($this->exactly(4))
             ->method('getParam')
-            ->withConsecutive(['message'], ['_store'], ['quote'])
-            ->willReturnOnConsecutiveCalls("Error Message", self::QUOTE_ID);
+            ->withConsecutive(['message'], ['_store'], ['quote'], ['orderId'])
+            ->willReturnOnConsecutiveCalls('Some message', $storeId, $encrypted, $encryptedOrderId);
 
-        $this->messageManager->expects($this->once())->method('addError')->willReturn($this->request);
+        $this->encryptorMock
+            ->expects($this->exactly(2))
+            ->method('decrypt')
+            ->withConsecutive([$encrypted], [$encryptedOrderId])
+            ->willReturnOnConsecutiveCalls($quoteId, $orderId);
 
-        $this->quote->expects($this->once())->method("getId")->willReturn(self::QUOTE_ID);
-        $this->quote->expects($this->once())->method("load")->willReturnSelf();
-        $this->quote->expects($this->once())->method("getReservedOrderId")->willReturn(self::RESERVED_ORDER_ID);
-
-        $this->orderFactory->expects($this->once())->method("create")->willReturn($this->order);
-        $this->order->expects($this->once())->method("loadByIncrementId")->with(self::RESERVED_ORDER_ID)->willReturnSelf();
-        $this->order->expects($this->once())->method("getId")->willReturn(self::RESERVED_ORDER_ID);
+        $this->quote->expects($this->once())->method("getId")->willReturn($quoteId);//self::QUOTE_ID
+        $this->quote->expects($this->once())->method("load")->with($quoteId)->willReturnSelf();//self::QUOTE_ID
 
         $this->recoverCartMock
             ->expects($this->once())
             ->method('setShouldCancelOrder')
             ->with(true)
+            ->willReturnSelf();
+        $this->recoverCartMock
+            ->expects($this->once())
+            ->method('setOrderId')
+            ->with($orderId)
             ->willReturnSelf();
         $this->recoverCartMock
             ->expects($this->once())